cran2deb: satisfy immediate R dependencies iff the requisite package has already been build.
Works by running apt-get update before each build, pulling in the
declared R dependencies from a local repository.

Still to do is actually read off a topological order and deal with
transitivity.  Otherwise, this appears to work fine where the only
dependencies are those in Depends or Imports of the R DESCRIPTION.

Also included is to use Imports as well as Depends for generating R
dependencies... this was necessary for, for example, stashR --- not
quite sure if this is technically correct, but it allows this package to
work.