cran2deb: cache availabile packages.  support cross-repo dependencies[1].  basic understanding of bundles (no building yet).
If an R package name is needed, and cannot be found in the available
packages, then try to resolve it into a bundle.  If this works, then
substitute the name of the bundle for the package name and procede. This
is enough to get dependency resolution working and R source packages
downloaded, but still to do is the generation of debian/ for bundles.

[1] Appears to be some problems during building of bioc packages -- even
though the package is called r-bioc-XXX, a directory r-cran-XXX is
expected by some part.  Suspect need to change generation of
debian/rules.