cran2deb: extra dependencies on command line; fix nasty bug in cross-repo dependencies.
accept with a common typo of version in License:.

bail out on SystemRequirements

correct a nasty bug: dependencies[r,]$name displayed like a string in R,
but was actually treated as a number; hence some cross-repo dependencies
did not work correctly since the wrong available[] entry was being used.

allow some extra dependencies to be specified on the command line.