licenses+db: delegate license acceptance to the database. add license management interface.
``License: file FOO'' support is nearly done; just need to plug bits
together on the acceptance side of things.  An SHA1 hash of the license
file is stored in the DB for matching.  Unsure how effective this will
be.  Might want to remove all whitespace prior to hashing.

needs testing; probably has a few bugs lurking. still need to work out
appropriate place for database so that R does not wipe it each time.
Perhaps it's time to delve into /var.

also a fix to r.dependency.closure --- previously incorrectly uses
levels() on the wrong type.