license: hashes of freeform licenses are stored in the database and these hashes used for auto-acceptance.
freeform licenses may be files, or may be the contents of the License:
field in the R DESCRIPTION.  such text is mapped to lower case and all
space characters are compressed and mapped to a single space.

a nicer interface for adding these freeform licenses is introduced.
after reviewing the license, its hash may be added as follows:

    $ cran2deb license
    license> add uroot gpl
    (maps hash of whatever freeform license uroot has to gpl)
    ...
    $ cran2deb build uroot
    (success is assured!)