cran2deb: use /var/cache/cran2deb as a permanent cache between installs.
Previously, when a new cran2deb was installed, cran2deb update would
re-generate the database and cache in their entirety, as well as lose
all previously generated .debs.  Instead they are stored outside the R
package heirarchy and so persist.

The cache is intended to contain everything that should be kept
in-memory for cran2deb (e.g., common data structures like the list of
all available packages), whilst the database is for all other data to be
stored on disk.