depends: add forced dependencies, separate run/build depends.
changes sufficient to make failures in rJava and RBGL external: openjdk
is not ready yet (latest crashes during configure), and
libboost-graph-dev 1.34.1-11 is not compatible with gcc 4.3 (see BioC
list archives).

also limit the frequency of pbuilder update and cache update.

note: database must be deleted and rebuilt with this patch.